requirements.txt
Django==1.11.3
djangorestframework==3.6.3
pytk==0.0.1.4
pytz==2017.2
